requirements.txt
pre-commit~=4.2.0
black~=24.4.2
isort~=5.13.2
flake8~=7.1.0
pillow~=11.1.0
